+++ src/StateSmithTest/spec/2/c/Spec2Sm.c
@@ -15,6 +15,7 @@ static void DECIDE_ev3(Spec2Sm* self);
 static void DECIDE_ev4(Spec2Sm* self);
 static void DECIDE_ev5(Spec2Sm* self);
 static void DECIDE_ev6(Spec2Sm* self);
+static void DECIDE_ev7(Spec2Sm* self);
 
 static void TEST1_DO_EVENT_TESTING_enter(Spec2Sm* self);
 static void TEST1_DO_EVENT_TESTING_exit(Spec2Sm* self);
@@ -142,6 +143,31 @@ static void TEST6_S1_ev1(Spec2Sm* self);
 static void TEST6_S2_enter(Spec2Sm* self);
 static void TEST6_S2_exit(Spec2Sm* self);
 
+static void TEST7_CHOICE_enter(Spec2Sm* self);
+static void TEST7_CHOICE_exit(Spec2Sm* self);
+
+static void TEST7_ROOT_enter(Spec2Sm* self);
+static void TEST7_ROOT_exit(Spec2Sm* self);
+static void TEST7_ROOT_ev4(Spec2Sm* self);
+static void TEST7_ROOT_ev5(Spec2Sm* self);
+
+static void TEST7_G_enter(Spec2Sm* self);
+static void TEST7_G_exit(Spec2Sm* self);
+static void TEST7_G_ev2(Spec2Sm* self);
+
+static void TEST7_G_S1_enter(Spec2Sm* self);
+static void TEST7_G_S1_exit(Spec2Sm* self);
+
+static void TEST7_G_S2_enter(Spec2Sm* self);
+static void TEST7_G_S2_exit(Spec2Sm* self);
+
+static void TEST7_G_S3_enter(Spec2Sm* self);
+static void TEST7_G_S3_exit(Spec2Sm* self);
+
+static void TEST7_S1_enter(Spec2Sm* self);
+static void TEST7_S1_exit(Spec2Sm* self);
+static void TEST7_S1_ev1(Spec2Sm* self);
+
 static void exit_up_to_state_handler(Spec2Sm* self, const Spec2Sm_Func desired_state_exit_handler);
 
 
@@ -240,6 +266,13 @@ const char* Spec2Sm_state_id_to_string(const enum Spec2Sm_StateId id)
         case Spec2Sm_StateId_TEST6_ROOT: return "TEST6_ROOT";
         case Spec2Sm_StateId_TEST6_S1: return "TEST6_S1";
         case Spec2Sm_StateId_TEST6_S2: return "TEST6_S2";
+        case Spec2Sm_StateId_TEST7_CHOICE: return "TEST7_CHOICE";
+        case Spec2Sm_StateId_TEST7_ROOT: return "TEST7_ROOT";
+        case Spec2Sm_StateId_TEST7_G: return "TEST7_G";
+        case Spec2Sm_StateId_TEST7_G_S1: return "TEST7_G_S1";
+        case Spec2Sm_StateId_TEST7_G_S2: return "TEST7_G_S2";
+        case Spec2Sm_StateId_TEST7_G_S3: return "TEST7_G_S3";
+        case Spec2Sm_StateId_TEST7_S1: return "TEST7_S1";
         default: return "?";
     }
 }
@@ -289,6 +322,7 @@ static void DECIDE_enter(Spec2Sm* self)
     self->current_event_handlers[Spec2Sm_EventId_EV4] = DECIDE_ev4;
     self->current_event_handlers[Spec2Sm_EventId_EV5] = DECIDE_ev5;
     self->current_event_handlers[Spec2Sm_EventId_EV6] = DECIDE_ev6;
+    self->current_event_handlers[Spec2Sm_EventId_EV7] = DECIDE_ev7;
     
     // DECIDE behavior
     // uml action: trace("Enter DECIDE.");
@@ -315,6 +349,7 @@ static void DECIDE_exit(Spec2Sm* self)
     self->current_event_handlers[Spec2Sm_EventId_EV4] = NULL;  // no ancestor listens to this event
     self->current_event_handlers[Spec2Sm_EventId_EV5] = NULL;  // no ancestor listens to this event
     self->current_event_handlers[Spec2Sm_EventId_EV6] = NULL;  // no ancestor listens to this event
+    self->current_event_handlers[Spec2Sm_EventId_EV7] = NULL;  // no ancestor listens to this event
 }
 
 static void DECIDE_ev1(Spec2Sm* self)
@@ -576,6 +611,53 @@ static void DECIDE_ev6(Spec2Sm* self)
                 return; // event processing immediately stops when a transition finishes. No other behaviors for this state are checked.
             } // end of behavior for TEST6_ROOT.InitialState
         } // end of behavior for TEST6_VARIABLES.InitialState
+    } // end of behavior for DECIDE
+}
+
+static void DECIDE_ev7(Spec2Sm* self)
+{
+    // No ancestor state handles `EV7` event.
+    
+    // DECIDE behavior
+    // uml guard: trace_guard("State DECIDE: check behavior `EV7 TransitionTo(TEST7_CHOICE)`.", true)
+    // uml action: trace("Transition action `` for DECIDE to TEST7_CHOICE.");
+    // uml transition target: TEST7_CHOICE
+    if (trace_guard("State DECIDE: check behavior `EV7 TransitionTo(TEST7_CHOICE)`.", true))
+    {
+        // Note: no `consume_event` variable possible here because of state transition. The event must be consumed.
+        // Optimize away while-exit-loop because we know that the active leaf state is DECIDE and it is exiting directly to its parent ROOT.
+        DECIDE_exit(self);
+        trace("Transition action `` for DECIDE to TEST7_CHOICE.");
+        
+        // Enter towards target
+        TEST7_CHOICE_enter(self);
+        
+        // TEST7_CHOICE.InitialState behavior
+        // uml action: trace("Transition action `` for TEST7_CHOICE.InitialState to TEST7_ROOT.");
+        // uml transition target: TEST7_ROOT
+        if (true)
+        {
+            trace("Transition action `` for TEST7_CHOICE.InitialState to TEST7_ROOT.");
+            
+            // Enter towards target
+            TEST7_ROOT_enter(self);
+            
+            // TEST7_ROOT.InitialState behavior
+            // uml action: trace("Transition action `` for TEST7_ROOT.InitialState to TEST7_S1.");
+            // uml transition target: TEST7_S1
+            if (true)
+            {
+                trace("Transition action `` for TEST7_ROOT.InitialState to TEST7_S1.");
+                
+                // Enter towards target
+                TEST7_S1_enter(self);
+                
+                // update state_id
+                self->state_id = Spec2Sm_StateId_TEST7_S1;
+                self->ancestor_event_handler = NULL;
+                return; // event processing immediately stops when a transition finishes. No other behaviors for this state are checked.
+            } // end of behavior for TEST7_ROOT.InitialState
+        } // end of behavior for TEST7_CHOICE.InitialState
     } // end of behavior for DECIDE
 }
 
@@ -2406,6 +2488,378 @@ static void TEST6_S2_exit(Spec2Sm* self)
     
     // adjust function pointers for this state's exit
     self->current_state_exit_handler = TEST6_ROOT_exit;
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_CHOICE
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_CHOICE_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_CHOICE_exit;
+    
+    // TEST7_CHOICE behavior
+    // uml action: trace("Enter TEST7_CHOICE.");
+    if (true)
+    {
+        trace("Enter TEST7_CHOICE.");
+    } // end of behavior for TEST7_CHOICE
+}
+
+static void TEST7_CHOICE_exit(Spec2Sm* self)
+{
+    // TEST7_CHOICE behavior
+    // uml action: trace("Exit TEST7_CHOICE.");
+    if (true)
+    {
+        trace("Exit TEST7_CHOICE.");
+    } // end of behavior for TEST7_CHOICE
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = ROOT_exit;
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_ROOT
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_ROOT_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_ROOT_exit;
+    self->current_event_handlers[Spec2Sm_EventId_EV4] = TEST7_ROOT_ev4;
+    self->current_event_handlers[Spec2Sm_EventId_EV5] = TEST7_ROOT_ev5;
+    
+    // TEST7_ROOT behavior
+    // uml action: trace("Enter TEST7_ROOT.");
+    if (true)
+    {
+        trace("Enter TEST7_ROOT.");
+    } // end of behavior for TEST7_ROOT
+    
+    // TEST7_ROOT behavior
+    // uml guard: trace_guard("State TEST7_ROOT: check behavior `enter / { clear_output(); }`.", true)
+    // uml action: clear_output();
+    if (trace_guard("State TEST7_ROOT: check behavior `enter / { clear_output(); }`.", true))
+    {
+        trace("IGNORE_OUTPUT_BEFORE_THIS");;
+    } // end of behavior for TEST7_ROOT
+}
+
+static void TEST7_ROOT_exit(Spec2Sm* self)
+{
+    // TEST7_ROOT behavior
+    // uml action: trace("Exit TEST7_ROOT.");
+    if (true)
+    {
+        trace("Exit TEST7_ROOT.");
+    } // end of behavior for TEST7_ROOT
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = TEST7_CHOICE_exit;
+    self->current_event_handlers[Spec2Sm_EventId_EV4] = NULL;  // no ancestor listens to this event
+    self->current_event_handlers[Spec2Sm_EventId_EV5] = NULL;  // no ancestor listens to this event
+}
+
+static void TEST7_ROOT_ev4(Spec2Sm* self)
+{
+    // No ancestor state handles `EV4` event.
+    
+    // TEST7_ROOT behavior
+    // uml guard: trace_guard("State TEST7_ROOT: check behavior `EV4 / { count--; }`.", true)
+    // uml action: count--;
+    if (trace_guard("State TEST7_ROOT: check behavior `EV4 / { count--; }`.", true))
+    {
+        // note: no ancestor consumes this event, but we output `bool consume_event` anyway because a user's design might rely on it.
+        bool consume_event = true; // events other than `do` are normally consumed by any event handler. Other event handlers in *this* state may still handle the event though.
+        (void)consume_event; // avoid un-used variable compiler warning. StateSmith cannot (yet) detect if behavior action code sets `consume_event`.
+        self->vars.count--;
+        
+        // No ancestor handles event. Ignore `consume_event` flag.
+    } // end of behavior for TEST7_ROOT
+}
+
+static void TEST7_ROOT_ev5(Spec2Sm* self)
+{
+    // No ancestor state handles `EV5` event.
+    
+    // TEST7_ROOT behavior
+    // uml guard: trace_guard("State TEST7_ROOT: check behavior `EV5 / { count++; }`.", true)
+    // uml action: count++;
+    if (trace_guard("State TEST7_ROOT: check behavior `EV5 / { count++; }`.", true))
+    {
+        // note: no ancestor consumes this event, but we output `bool consume_event` anyway because a user's design might rely on it.
+        bool consume_event = true; // events other than `do` are normally consumed by any event handler. Other event handlers in *this* state may still handle the event though.
+        (void)consume_event; // avoid un-used variable compiler warning. StateSmith cannot (yet) detect if behavior action code sets `consume_event`.
+        self->vars.count++;
+        
+        // No ancestor handles event. Ignore `consume_event` flag.
+    } // end of behavior for TEST7_ROOT
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_G
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_G_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_G_exit;
+    self->current_event_handlers[Spec2Sm_EventId_EV2] = TEST7_G_ev2;
+    
+    // TEST7_G behavior
+    // uml action: trace("Enter TEST7_G.");
+    if (true)
+    {
+        trace("Enter TEST7_G.");
+    } // end of behavior for TEST7_G
+}
+
+static void TEST7_G_exit(Spec2Sm* self)
+{
+    // TEST7_G behavior
+    // uml action: trace("Exit TEST7_G.");
+    if (true)
+    {
+        trace("Exit TEST7_G.");
+    } // end of behavior for TEST7_G
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = TEST7_ROOT_exit;
+    self->current_event_handlers[Spec2Sm_EventId_EV2] = NULL;  // no ancestor listens to this event
+}
+
+static void TEST7_G_ev2(Spec2Sm* self)
+{
+    // No ancestor state handles `EV2` event.
+    
+    // TEST7_G behavior
+    // uml guard: trace_guard("State TEST7_G: check behavior `EV2 TransitionTo(TEST7_ROOT.InitialState)`.", true)
+    // uml action: trace("Transition action `` for TEST7_G to TEST7_ROOT.InitialState.");
+    // uml transition target: TEST7_ROOT.InitialState
+    if (trace_guard("State TEST7_G: check behavior `EV2 TransitionTo(TEST7_ROOT.InitialState)`.", true))
+    {
+        // Note: no `consume_event` variable possible here because of state transition. The event must be consumed.
+        // At this point, StateSmith doesn't know what the active leaf state is. It could be TEST7_G or one of its sub states.
+        exit_up_to_state_handler(self, TEST7_ROOT_exit);  // Exit until we reach TEST7_ROOT state.
+        trace("Transition action `` for TEST7_G to TEST7_ROOT.InitialState.");
+        
+        // Enter towards target
+        
+        // TEST7_ROOT.InitialState behavior
+        // uml action: trace("Transition action `` for TEST7_ROOT.InitialState to TEST7_S1.");
+        // uml transition target: TEST7_S1
+        if (true)
+        {
+            trace("Transition action `` for TEST7_ROOT.InitialState to TEST7_S1.");
+            
+            // Enter towards target
+            TEST7_S1_enter(self);
+            
+            // update state_id
+            self->state_id = Spec2Sm_StateId_TEST7_S1;
+            self->ancestor_event_handler = NULL;
+            return; // event processing immediately stops when a transition finishes. No other behaviors for this state are checked.
+        } // end of behavior for TEST7_ROOT.InitialState
+    } // end of behavior for TEST7_G
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_G_S1
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_G_S1_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_G_S1_exit;
+    
+    // TEST7_G_S1 behavior
+    // uml action: trace("Enter TEST7_G_S1.");
+    if (true)
+    {
+        trace("Enter TEST7_G_S1.");
+    } // end of behavior for TEST7_G_S1
+}
+
+static void TEST7_G_S1_exit(Spec2Sm* self)
+{
+    // TEST7_G_S1 behavior
+    // uml action: trace("Exit TEST7_G_S1.");
+    if (true)
+    {
+        trace("Exit TEST7_G_S1.");
+    } // end of behavior for TEST7_G_S1
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = TEST7_G_exit;
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_G_S2
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_G_S2_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_G_S2_exit;
+    
+    // TEST7_G_S2 behavior
+    // uml action: trace("Enter TEST7_G_S2.");
+    if (true)
+    {
+        trace("Enter TEST7_G_S2.");
+    } // end of behavior for TEST7_G_S2
+}
+
+static void TEST7_G_S2_exit(Spec2Sm* self)
+{
+    // TEST7_G_S2 behavior
+    // uml action: trace("Exit TEST7_G_S2.");
+    if (true)
+    {
+        trace("Exit TEST7_G_S2.");
+    } // end of behavior for TEST7_G_S2
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = TEST7_G_exit;
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_G_S3
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_G_S3_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_G_S3_exit;
+    
+    // TEST7_G_S3 behavior
+    // uml action: trace("Enter TEST7_G_S3.");
+    if (true)
+    {
+        trace("Enter TEST7_G_S3.");
+    } // end of behavior for TEST7_G_S3
+}
+
+static void TEST7_G_S3_exit(Spec2Sm* self)
+{
+    // TEST7_G_S3 behavior
+    // uml action: trace("Exit TEST7_G_S3.");
+    if (true)
+    {
+        trace("Exit TEST7_G_S3.");
+    } // end of behavior for TEST7_G_S3
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = TEST7_G_exit;
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// event handlers for state TEST7_S1
+////////////////////////////////////////////////////////////////////////////////
+
+static void TEST7_S1_enter(Spec2Sm* self)
+{
+    // setup trigger/event handlers
+    self->current_state_exit_handler = TEST7_S1_exit;
+    self->current_event_handlers[Spec2Sm_EventId_EV1] = TEST7_S1_ev1;
+    
+    // TEST7_S1 behavior
+    // uml action: trace("Enter TEST7_S1.");
+    if (true)
+    {
+        trace("Enter TEST7_S1.");
+    } // end of behavior for TEST7_S1
+}
+
+static void TEST7_S1_exit(Spec2Sm* self)
+{
+    // TEST7_S1 behavior
+    // uml action: trace("Exit TEST7_S1.");
+    if (true)
+    {
+        trace("Exit TEST7_S1.");
+    } // end of behavior for TEST7_S1
+    
+    // adjust function pointers for this state's exit
+    self->current_state_exit_handler = TEST7_ROOT_exit;
+    self->current_event_handlers[Spec2Sm_EventId_EV1] = NULL;  // no ancestor listens to this event
+}
+
+static void TEST7_S1_ev1(Spec2Sm* self)
+{
+    // No ancestor state handles `EV1` event.
+    
+    // TEST7_S1 behavior
+    // uml guard: trace_guard("State TEST7_S1: check behavior `EV1 TransitionTo(TEST7_G)`.", true)
+    // uml action: trace("Transition action `` for TEST7_S1 to TEST7_G.");
+    // uml transition target: TEST7_G
+    if (trace_guard("State TEST7_S1: check behavior `EV1 TransitionTo(TEST7_G)`.", true))
+    {
+        // Note: no `consume_event` variable possible here because of state transition. The event must be consumed.
+        // Optimize away while-exit-loop because we know that the active leaf state is TEST7_S1 and it is exiting directly to its parent TEST7_ROOT.
+        TEST7_S1_exit(self);
+        trace("Transition action `` for TEST7_S1 to TEST7_G.");
+        
+        // Enter towards target
+        TEST7_G_enter(self);
+        
+        // TEST7_G.InitialState behavior
+        // uml guard: count == 1
+        // uml action: trace("Transition action `` for TEST7_G.InitialState to TEST7_G_S1.");
+        // uml transition target: TEST7_G_S1
+        if (self->vars.count == 1)
+        {
+            trace("Transition action `` for TEST7_G.InitialState to TEST7_G_S1.");
+            
+            // Enter towards target
+            TEST7_G_S1_enter(self);
+            
+            // update state_id
+            self->state_id = Spec2Sm_StateId_TEST7_G_S1;
+            self->ancestor_event_handler = NULL;
+            return; // event processing immediately stops when a transition finishes. No other behaviors for this state are checked.
+        } // end of behavior for TEST7_G.InitialState
+        // TEST7_G.InitialState behavior
+        // uml guard: count == 2
+        // uml action: trace("Transition action `` for TEST7_G.InitialState to TEST7_G_S2.");
+        // uml transition target: TEST7_G_S2
+        if (self->vars.count == 2)
+        {
+            trace("Transition action `` for TEST7_G.InitialState to TEST7_G_S2.");
+            
+            // Enter towards target
+            TEST7_G_S2_enter(self);
+            
+            // update state_id
+            self->state_id = Spec2Sm_StateId_TEST7_G_S2;
+            self->ancestor_event_handler = NULL;
+            return; // event processing immediately stops when a transition finishes. No other behaviors for this state are checked.
+        } // end of behavior for TEST7_G.InitialState
+        // TEST7_G.InitialState behavior
+        // uml action: trace("Transition action `` for TEST7_G.InitialState to TEST7_G_S3.");
+        // uml transition target: TEST7_G_S3
+        if (true)
+        {
+            trace("Transition action `` for TEST7_G.InitialState to TEST7_G_S3.");
+            
+            // Enter towards target
+            TEST7_G_S3_enter(self);
+            
+            // update state_id
+            self->state_id = Spec2Sm_StateId_TEST7_G_S3;
+            self->ancestor_event_handler = NULL;
+            return; // event processing immediately stops when a transition finishes. No other behaviors for this state are checked.
+        } // end of behavior for TEST7_G.InitialState
+    } // end of behavior for TEST7_S1
 }