+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support initial/entry/exit multiple https://github.com/StateSmith/StateSmith/issues/40
-
+  - Allow multiple transitions from initial state.
+  - Initial state must have a default transition (always true).
+  - Allow incoming transitions to initial state.